requirements.txt
mysql_connector_repackaged==0.3.1
numpy==2.4.4
python-dotenv==1.2.2
scipy==1.17.1
